improved api structure
