+++ src/Animation.h
@@ -1,22 +1,25 @@
+#pragma once
 #include "imgui.h"
 #include "math.h"
 
 class Animation{
 	float mTick{0.0f};
 	float mDuration{0.0f};
+	bool mHasCompleted=false;
 
-	//Animation Function
 	inline float EaseOutQuadraticFn(float t) { return 1.0f - pow(1.0f - t, 4);}
 
-	bool mHasCompleted=false;
+
+
 public:
 	bool hasStarted=false;
+
+
 	Animation(float duration=0.5f):mDuration(duration){}
 
-	void start(){
-		mHasCompleted=false;
-		hasStarted=true;
-	}
+	void SetDuration(float dur){ mDuration=dur;}
+
+	void start(){ mHasCompleted=false; hasStarted=true;}
 
 	float update(){
 		if(mHasCompleted) return 1.0f;
