+++ src/main.cpp
@@ -6,13 +6,12 @@
 #include <shellapi.h>
 #include <sstream>
 #include <string>
-#include <unordered_map>
 #include <winuser.h>
 #define STB_IMAGE_IMPLEMENTATION
-#include "TextEditor.h"
 #include "stb_img.h"
 #include <codecvt>
-#include "DirectoryHandler.h"
+#include "CoreSystem.h"
+#include "TextEditor.h"
 
 std::string exec(const char* cmd) {
 	// bool status = !std::system(cmd);
@@ -40,15 +39,7 @@ std::string exec(const char* cmd) {
 #define HEIGHT 600
 
 Editor editor;
-struct Entity{
-	std::string filename;
-	std::string path;
-	bool is_directory=false;
-	bool is_explored=false;
-};
-
-std::unordered_map<std::string,std::vector<Entity>> mDirectoryData;
-
+CoreSystem coreSystem;
 
 
 
@@ -56,7 +47,7 @@ void draw(GLFWwindow* window, ImGuiIO& io);
 
 void framebuffer_size_callback(GLFWwindow* window, int width, int height)
 {
-	editor.reCalculateBounds = true;
+	coreSystem.GetTextEditor()->RecalculateBounds();
 	glViewport(0, 0, width, height);
 	draw(window, ImGui::GetIO());
 }
@@ -66,120 +57,14 @@ void drop_callback(GLFWwindow* window, int count, const char** paths)
 	for (int i = 0; i < count; i++) {
 		if (std::filesystem::is_directory(paths[i])) {
 			GL_INFO("Folder: {}", paths[i]);
-			editor.AddFolder(paths[i]);
+			coreSystem.GetFileNavigation()->AddFolder(paths[i]);
 		} else {
 			GL_INFO("File: {}", paths[i]);
-			editor.LoadFile(paths[i]);
+			coreSystem.GetTextEditor()->LoadFile(paths[i]);
 		}
 	}
 }
 
-inline void ShowContextMenu(std::string& path,std::string& name,bool isFolder=true){
-    static int selected=-1;
-
-    if (ImGui::BeginPopupContextItem(name.c_str()))
-    {
-	    const char* options[] = {ICON_FA_CARET_RIGHT"  New File",ICON_FA_CARET_RIGHT"  Rename",ICON_FA_CARET_RIGHT"  Open Folder",ICON_FA_CARET_RIGHT"  Open Terminal",ICON_FA_CARET_RIGHT"  New Folder",ICON_FA_CARET_RIGHT"  Delete Folder"};
-	    if(!isFolder){
-	    	options[5]=ICON_FA_CARET_RIGHT"  Delete File";
-	    }else{
-	    	options[5]=ICON_FA_CARET_RIGHT"  Delete Folder";
-	    }
-        for (int i = 0; i < IM_ARRAYSIZE(options); i++){
-        	if(i==4) ImGui::Separator();
-            if (ImGui::Selectable(options[i])){
-            	selected = i;
-                std::string fpath=path;
-            	switch(selected){
-                	case 0:
-                		GL_INFO("New File");
-                		GL_INFO("{}/{}",path,name);
-                		break;
-                	case 1:
-                		GL_INFO("Rename");
-                		break;
-                	case 2:
-						if(std::filesystem::is_directory(path))
-							fpath = (std::filesystem::path(path)/name).generic_string();
-						DirectoryHandler::OpenExplorer(fpath);
-                		break;
-                	case 3:
-						if(std::filesystem::is_directory(path))
-							fpath = (std::filesystem::path(path)/name).generic_string();
-						DirectoryHandler::StartTerminal(fpath);
-                		break;
-                	case 4:
-                		GL_INFO("New Folder");
-                		break;
-                	case 5:
-                		GL_INFO("Delete Folder");
-                		break;
-            	}
-
-            }
-        }
-        ImGui::EndPopup();
-    }
-}
-
-void RenderFolderItems(std::string path,bool isRoot=false){
-	if(isRoot){
-		std::string folderName=std::filesystem::path(path).filename().generic_string();
-		ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing,ImVec2(6.0f,2.0f));
-		ImGui::PushStyleVar(ImGuiStyleVar_FramePadding,ImVec2(4.0f,2.0f));
-		if(ImGui::TreeNodeEx(folderName.c_str(),ImGuiTreeNodeFlags_SpanFullWidth | ImGuiTreeNodeFlags_DefaultOpen)){
-    		RenderFolderItems(path);
-    		ImGui::TreePop();
-    	}
-		ShowContextMenu(path,folderName);
-    	ImGui::PopStyleVar(2);
-    	return;
-	}
-	if(mDirectoryData.empty()  || mDirectoryData.find(path)==mDirectoryData.end()){
-		auto& entities=mDirectoryData[path];
-		for(const auto& entity:std::filesystem::directory_iterator(path)) entities.push_back({
-			entity.path().filename().generic_string(),
-			entity.path().generic_string(),
-			entity.is_directory(),
-			false
-		});
-		std::stable_partition(entities.begin(), entities.end(), [](const auto& entity) { return entity.is_directory; });
-	}
-	auto& entities=mDirectoryData[path];
-	if(entities.empty()) return;
-	std::stringstream oss;
-	for(Entity& item:entities){
-		const char* icon = item.is_directory ? item.is_explored ? ICON_FA_FOLDER_OPEN : ICON_FA_FOLDER : ICON_FA_FILE;
-		oss << icon << " " << item.filename.c_str();
-		if(item.is_directory) {
-			if(ImGui::TreeNodeEx(oss.str().c_str(),ImGuiTreeNodeFlags_Framed | ImGuiTreeNodeFlags_SpanFullWidth | ImGuiTreeNodeFlags_Selected)){
-				std::stringstream wss;
-				ShowContextMenu(path,item.filename); //Explored Folder
-				wss << path << "\\" << item.filename.c_str();
-				RenderFolderItems(wss.str(),false);
-				ImGui::TreePop();
-			}
-			ShowContextMenu(path,item.filename); // UnExplored Folder
-		}
-		else{
-			ImGui::PushFont(ImGui::GetIO().Fonts->Fonts[1]);
-			ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing,ImVec2(6.0f,8.0f));
-			ImGui::PushStyleColor(ImGuiCol_Header,ImGui::GetStyle().Colors[ImGuiCol_SliderGrabActive]);
-			ImGui::PushStyleColor(ImGuiCol_HeaderHovered,ImGui::GetStyle().Colors[ImGuiCol_SliderGrab]);
-			ImGui::SetCursorPosX(ImGui::GetCursorPosX()+28.0f);
-			if(ImGui::Selectable(oss.str().c_str(),item.is_explored,ImGuiSelectableFlags_SpanAvailWidth|ImGuiSelectableFlags_SelectOnClick)){
-				for(Entity& en:entities) en.is_explored=false;
-				if(!item.is_explored) item.is_explored=true;
-				editor.LoadFile(item.path.c_str());
-			}
-			ImGui::PopFont();
-			ShowContextMenu(path,item.filename,false);
-			ImGui::PopStyleVar();
-			ImGui::PopStyleColor(2);
-		}
-		oss.str("");
-	}
-}
 
 void draw(GLFWwindow* window, ImGuiIO& io)
 {
@@ -189,258 +74,7 @@ void draw(GLFWwindow* window, ImGuiIO& io)
 	ImGui::NewFrame();
 	glfwSetDropCallback(window, drop_callback);
 
-	static ImGuiDockNodeFlags dockspace_flags = ImGuiDockNodeFlags_PassthruCentralNode;
-	static ImGuiWindowFlags window_flags = ImGuiWindowFlags_None | ImGuiWindowFlags_MenuBar | ImGuiWindowFlags_NoDocking;
-
-	static bool isPanelOpen=false;
-	static const float pannelSize=38.0f;
-	static const float statusSize=22.0f;
-	const ImGuiViewport* viewport = ImGui::GetMainViewport();
-	ImGui::SetNextWindowPos(viewport->WorkPos);
-	ImVec2 size(viewport->WorkSize.x,isPanelOpen ? viewport->WorkSize.y-statusSize - pannelSize : viewport->WorkSize.y-statusSize);
-	ImGui::SetNextWindowSize(size);
-	ImGui::SetNextWindowViewport(viewport->ID);
-	ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding, 0.0f);
-	ImGui::PushStyleVar(ImGuiStyleVar_WindowBorderSize, 0.0f);
-	window_flags |= ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove;
-	window_flags |= ImGuiWindowFlags_NoBringToFrontOnFocus | ImGuiWindowFlags_NoNavFocus;
-    if (dockspace_flags & ImGuiDockNodeFlags_PassthruCentralNode)
-    	window_flags |= ImGuiWindowFlags_NoBackground;
-
-	ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, ImVec2(0.0f, 0.0f));
-	ImGui::Begin("Container",nullptr, window_flags);
-	ImGui::PopStyleVar(3);
-
-	if (io.ConfigFlags & ImGuiConfigFlags_DockingEnable){
-        ImGuiID dockspace_id = ImGui::GetID("DDockSpace");
-        ImGui::DockSpace(dockspace_id, ImVec2(0.0f, 0.0f), dockspace_flags);
-
-		static bool isFirst=true;
-		if(isFirst){
-			isFirst=false;
-			ImGui::DockBuilderRemoveNode(dockspace_id); // clear any previous layout
-			ImGui::DockBuilderAddNode(dockspace_id,dockspace_flags | ImGuiDockNodeFlags_DockSpace);
-			ImGui::DockBuilderSetNodeSize(dockspace_id, size);
-
-			auto dock_id_left = ImGui::DockBuilderSplitNode(dockspace_id, ImGuiDir_Left, 0.3f, nullptr, &dockspace_id);
-			// auto dock_id_right = ImGui::DockBuilderSplitNode(dockspace_id, ImGuiDir_Right, 0.7f, nullptr, &dockspace_id);
-			ImGui::DockBuilderDockWindow("Project Directory", dock_id_left);
-			ImGui::DockBuilderDockWindow("Editor", dockspace_id);
-			ImGui::DockBuilderFinish(dockspace_id);
-		}
-	}
-
-	if (ImGui::BeginMenuBar()) {
-		if (ImGui::BeginMenu("File")) {
-			ImGui::MenuItem("New File");
-
-			if(ImGui::MenuItem("Open File"))
-				editor.LoadFile(SelectFile().c_str());
-
-			if(ImGui::MenuItem("Open Folder")){
-				std::string path=SelectFolder();
-				if(!path.empty()) editor.AddFolder(path.c_str());
-			}
-			
-			ImGui::EndMenu();
-		}
-
-		if (ImGui::BeginMenu("Edit")) {
-			ImGui::MenuItem("Cut");
-			ImGui::MenuItem("Copy");
-			ImGui::MenuItem("Paste");
-            ImGui::EndMenu();
-		}
-		ImGui::EndMenuBar();
-	}
-
-	ImGui::End();
-	#ifdef GL_DEBUG
-	static bool show_demo=true;
-	ImGui::ShowDemoWindow(&show_demo);
-
-
-	ImGui::Begin("Project");
-
-
-	static int LineSpacing = 10.0f;
-	if (ImGui::SliderInt("LineSpacing", &LineSpacing, 0, 20)) {
-		editor.setLineSpacing(LineSpacing);
-	}
-
-    ImGui::Spacing();
-	ImGui::Text("PositionY:%.2f", ImGui::GetMousePos().y);
-    ImGui::Spacing();
-    ImGui::Text("mCursorPosition: X:%d  Y:%d",editor.GetEditorState()->mCursorPosition.mColumn,editor.GetEditorState()->mCursorPosition.mLine);
-    ImGui::Text("mSelectionStart: X:%d  Y:%d",editor.GetEditorState()->mSelectionStart.mColumn,editor.GetEditorState()->mSelectionStart.mLine);
-    ImGui::Text("mSelectionEnd:   X:%d  Y:%d",editor.GetEditorState()->mSelectionEnd.mColumn,editor.GetEditorState()->mSelectionEnd.mLine);
-
-
-    ImGui::Spacing();
-    ImGui::Spacing();
-    static std::string mode;
-    ImColor color;
-    switch(editor.GetSelectionMode()){
-	    case 0: 
-	    	mode="Normal";
-	    	color=ImColor(50,206,187,255);
-	    	break;
-	    case 1: 
-	    	mode="Word";
-	    	color=ImColor(233,196,106,255);
-	    	break;
-	    case 2: 
-	    	mode="Line";
-	    	color=ImColor(231,111,81,255);
-	    	break;
-    }
-    ImGui::Text("SelectionMode: ");
-    ImGui::SameLine();
-    ImGui::PushStyleColor(ImGuiCol_Text,color.Value);
-    ImGui::Text("%s", mode.c_str());
-    ImGui::PopStyleColor();
-
-
-
-    ImGui::Spacing();
-    ImGui::Spacing();
-    static int v{1};
-    ImGui::Text("Goto Line: "); ImGui::SameLine(); 
-    if(ImGui::InputInt("##ScrollToLine", &v,1,100,ImGuiInputTextFlags_EnterReturnsTrue))
-        editor.ScrollToLineNumber(v);
-
-    if(ImGui::Button("Select File")) SelectFile();
-
-    if(ImGui::Button("Select Files")){
-    	auto files=SelectFiles();
-    	for(auto& file:files)
-    		std::wcout << file << std::endl;
-    }
-
-	ImGui::End();
-	#endif
-
-	static float s_width=250.0f;
-	static bool is_opening=true;
-	static Animation side_bar;
-
-	if(is_opening){
-		ImGui::PushStyleColor(ImGuiCol_WindowBg,ImVec4(0.067f,0.075f,0.078f,1.000f));
-		ImGui::SetNextWindowSize(ImVec2{s_width,-1.0f},ImGuiCond_Once);
-		ImGui::PushStyleColor(ImGuiCol_Text,ImVec4(0.764,0.764,0.764,1.000));
-		ImGui::Begin("Project Directory");
-	    	auto& folders=editor.GetFolders();
-	    	for(const auto& folder:folders)
-	    		RenderFolderItems(folder,true);
-		ImGui::End();
-		ImGui::PopStyleColor(2);
-	}
-
-
-	editor.Render();
-
-	ImGui::PushStyleVar(ImGuiStyleVar_WindowBorderSize,0.0f);
-	if(isPanelOpen){
-
-		ImGui::SetNextWindowPos(ImVec2(0,size.y));
-		ImGui::SetNextWindowSize(ImVec2(viewport->WorkSize.x,pannelSize));
-		ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding,ImVec2(4.0f,0.0f));
-		ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding,0.0f);
-		ImGui::PushStyleColor(ImGuiCol_WindowBg,ImGui::GetStyle().Colors[ImGuiCol_TitleBg]);
-		ImGui::PushFont(io.Fonts->Fonts[1]);
-		ImGui::Begin("InputBar",0,ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoDocking | ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoResize );
-
-			const char* prompt_text="File Name:";
-			const ImVec2 txt_dim=ImGui::CalcTextSize(prompt_text);
-			const float y=(ImGui::GetWindowHeight()-txt_dim.y)*0.5f;
-
-			ImGui::PushFont(io.Fonts->Fonts[0]);
-			ImGui::SetCursorPosY(y-ImGui::GetStyle().FramePadding.y*2);
-			ImGui::Text("%s", prompt_text);
-
-
-			ImGui::SetCursorPos({txt_dim.x,5.0f});
-			static char buff[256]="Placeholder";
-			ImGui::PushStyleVar(ImGuiStyleVar_FramePadding,ImVec2(4.0f,6.0f));
-			ImGui::PushStyleColor(ImGuiCol_FrameBg,ImGui::GetStyle().Colors[ImGuiCol_MenuBarBg]);
-			ImGui::PushItemWidth(viewport->WorkSize.x-txt_dim.x-5.0f);
-			if(ImGui::InputText("##InputBar",buff,IM_ARRAYSIZE(buff),ImGuiInputTextFlags_EnterReturnsTrue)){
-				GL_INFO("Submit");
-			}
-			ImGui::PopStyleVar();
-			ImGui::PopStyleColor();
-			ImGui::PopFont();
-			if(ImGui::IsKeyPressed(ImGuiKey_Escape)){
-				isPanelOpen=false;
-			}
-		ImGui::End();
-		ImGui::PopFont();
-		ImGui::PopStyleColor();
-		ImGui::PopStyleVar(2);
-		
-	}
-
-
-
-
-	ImGui::SetNextWindowPos(ImVec2(0,isPanelOpen ? size.y+pannelSize : size.y));
-	ImGui::SetNextWindowSize(ImVec2(viewport->WorkSize.x,statusSize));
-	ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding,ImVec2(4.0f,0.0f));
-	ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding,0.0f);
-	ImGui::PushStyleColor(ImGuiCol_WindowBg,ImGui::GetStyle().Colors[ImGuiCol_TitleBg]);
-	ImGui::PushFont(io.Fonts->Fonts[1]);
-	ImGui::Begin("Status Bar",0,ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoDocking | ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoResize );
-		ImGui::PushFont(io.Fonts->Fonts[0]);
-		if(ImGui::Button(ICON_FA_BARS)){
-			is_opening=!is_opening;
-			isPanelOpen=true;
-		}
-		ImGui::PopFont();
-		ImGui::SameLine();
-		ImGui::SetCursorPosY(ImGui::GetCursorPosY()+2.0f);
-		ImGui::Text("Line:%d",editor.GetEditorState()->mCursorPosition.mLine+1);
-		ImGui::SameLine();
-		ImGui::SetCursorPosY(ImGui::GetCursorPosY()+2.0f);
-		ImGui::Text("Column:%d",editor.GetEditorState()->mCursorPosition.mColumn+1);
-
-		static Animation aFileSave(2.0f);
-
-		if(editor.isFileSaving){
-			editor.isFileSaving=false;
-			aFileSave.start();
-		}
-		if(aFileSave.hasStarted){
-			ImGui::SameLine();
-			ImGui::SetCursorPosY(ImGui::GetCursorPosY()+2.0f);
-			ImGui::TextColored(ImVec4(0.165,0.616,0.561,1.000),"Saved:%s",editor.GetCurrentFilePath().c_str());
-			aFileSave.update();
-		}
-
-
-		static bool branchLoaded =false;
-		static std::string branch;
-		if(!branchLoaded){
-			// branch=exec("git rev-parse --abbrev-ref HEAD > git.txt");
-			branchLoaded=true;
-		}
-		if(!branch.empty()){
-			ImGui::PushFont(io.Fonts->Fonts[0]);
-			ImGui::SameLine(ImGui::GetWindowWidth()-60.0f);
-			ImGui::SetCursorPosY(ImGui::GetCursorPosY());
-			ImGui::Text("%s %s",ICON_FA_CODE_BRANCH,branch.c_str());
-			ImGui::PopFont();
-		}
-
-		// FileType
-		float width=ImGui::CalcTextSize(editor.fileType.c_str()).x;
-		ImGui::SameLine(ImGui::GetWindowWidth()-width-10.0f);
-		ImGui::SetCursorPosY(ImGui::GetCursorPosY()+2.0f);
-		ImGui::Text("%s", editor.fileType.c_str());
-
-	ImGui::End();
-	ImGui::PopFont();
-	ImGui::PopStyleColor();
-	ImGui::PopStyleVar(3);
+	coreSystem.Render();
 
 	ImGui::Render();
 	int display_w, display_h;
@@ -470,10 +104,10 @@ void HandleArguments(int argc,char* argv[]){
         if (fs::exists(path)) {
             if (fs::is_regular_file(path)) {
             	GL_INFO("FILE:{}",path.generic_string());
-                editor.LoadFile(path.generic_string().c_str());
+                coreSystem.GetTextEditor()->LoadFile(path.generic_string().c_str());
             } else if (fs::is_directory(path)) {
             	GL_INFO("FOLDER:{}",path.generic_string());
-                editor.AddFolder(path.generic_string());
+                coreSystem.GetFileNavigation()->AddFolder(path.generic_string());
             } else {
             	ShowErrorMessage("Invalid File/Folder Selected");
             }
@@ -490,9 +124,10 @@ int main(int argc,char* argv[])
 	if(argc > 1) HandleArguments(argc,argv);
 
 	GLFWwindow* window;
-#ifdef GL_DEBUG
-	OpenGL::Log::Init();
-#endif
+	
+	#ifdef GL_DEBUG
+		OpenGL::Log::Init();
+	#endif
 
 	if (!glfwInit()) return -1;
 
@@ -564,8 +199,8 @@ int main(int argc,char* argv[])
 
     StyleColorDarkness();
 
-	editor.LoadFile("D:/Projects/c++/txedit/src/TextEditor.cpp");
-	editor.AddFolder("D:/Projects/c++/txedit");
+	coreSystem.GetTextEditor()->LoadFile("D:/Projects/c++/txedit/src/TextEditor.cpp");
+	coreSystem.GetFileNavigation()->AddFolder("D:/Projects/c++/txedit");
 
 	while (!glfwWindowShouldClose(window)) {
 		draw(window, io);
